Add applicative def for stream
Note app2 is broken because concatMap is. Also shook out some issues
where type needed to decay rather than being passed through .

Value&& in a deduced context can be an 'unusual' type. Deal with it
properly.

This showed up in the partial specialization of the make<ConsStream>, so
the underlying Make class would get Value deduced oddly. As a reference,
or const reference.